bdk datasource + wallet repository + removal of wallet manager service
