+++ lib/core/blockchain/data/repository/liquid_blockchain_repository_impl.dart
@@ -5,7 +5,7 @@ import 'package:bb_mobile/core/blockchain/domain/repositories/liquid_blockchain_
 import 'package:bb_mobile/core/electrum/data/datasources/electrum_server_storage_datasource.dart';
 import 'package:bb_mobile/core/electrum/data/models/electrum_server_model.dart';
 import 'package:bb_mobile/core/electrum/domain/entity/electrum_server.dart';
-import 'package:bb_mobile/core/wallet/domain/entity/wallet_metadata.dart';
+import 'package:bb_mobile/core/wallet/domain/entity/wallet.dart';
 
 class LiquidBlockchainRepositoryImpl implements LiquidBlockchainRepository {
   final LiquidBlockchainDatasource _blockchain;
