+++ lib/core/recoverbull/domain/entity/recoverbull_wallet.dart
@@ -1,4 +1,4 @@
-import 'package:bb_mobile/core/wallet/domain/entity/wallet_metadata.dart';
+import 'package:bb_mobile/core/wallet/domain/entity/wallet.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 
 part 'recoverbull_wallet.freezed.dart';
