+++ lib/core/wallet/data/datasources/lwk_wallet_datasource.dart
@@ -1,0 +1,1 @@
+class LwkWalletDatasource {}
