+++ lib/core/wallet/data/models/balance_model.dart
@@ -1,15 +1,15 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
 
-part 'balance.freezed.dart';
+part 'balance_model.freezed.dart';
 
 @freezed
-class Balance with _$Balance {
-  const factory Balance({
+class BalanceModel with _$BalanceModel {
+  const factory BalanceModel({
     required BigInt immatureSat,
     required BigInt trustedPendingSat,
     required BigInt untrustedPendingSat,
     required BigInt confirmedSat,
     required BigInt spendableSat,
     required BigInt totalSat,
-  }) = _Balance;
+  }) = _BalanceModel;
 }
