+++ lib/core/wallet/data/models/tx_input_model.dart
@@ -1,16 +1,15 @@
 import 'dart:typed_data';
 
-
 import 'package:freezed_annotation/freezed_annotation.dart';
 
-part 'tx_input.freezed.dart';
+part 'tx_input_model.freezed.dart';
 
 @freezed
-sealed class TxInput with _$TxInput {
-  const factory TxInput({
+sealed class TxInputModel with _$TxInputModel {
+  const factory TxInputModel({
     required String txId,
     required int vout,
     required Uint8List scriptPubkey,
-  }) = _TxInput;
-  const TxInput._();
+  }) = _TxInputModel;
+  const TxInputModel._();
 }
