+++ lib/core/wallet/data/models/utxo_model.dart
@@ -2,16 +2,15 @@ import 'dart:typed_data';
 
 import 'package:freezed_annotation/freezed_annotation.dart';
 
-part 'utxo.freezed.dart';
+part 'utxo_model.freezed.dart';
 
 @freezed
-sealed class Utxo with _$Utxo {
-  const factory Utxo({
+sealed class UtxoModel with _$UtxoModel {
+  const factory UtxoModel({
     required String txId,
     required int vout,
     required BigInt value,
     required Uint8List scriptPubkey,
-    String? label,
-  }) = _Utxo;
-  const Utxo._();
+  }) = _UtxoModel;
+  const UtxoModel._();
 }
