--- lib/core/wallet/data/repository/wallet_metadata_repository_impl.dart
@@ -1,83 +1,0 @@
-import 'package:bb_mobile/core/seed/domain/entity/seed.dart';
-import 'package:bb_mobile/core/wallet/data/datasources/wallet_metadata_datasource.dart';
-import 'package:bb_mobile/core/wallet/data/models/wallet_metadata_model.dart';
-import 'package:bb_mobile/core/wallet/domain/entity/wallet_metadata.dart';
-import 'package:bb_mobile/core/wallet/domain/repositories/wallet_metadata_repository.dart';
-
-class WalletMetadataRepositoryImpl implements WalletMetadataRepository {
-  final WalletMetadataDatasource _source;
-
-  WalletMetadataRepositoryImpl({required WalletMetadataDatasource source})
-      : _source = source;
-
-  @override
-  Future<WalletMetadata> deriveFromSeed({
-    required Seed seed,
-    required Network network,
-    required ScriptType scriptType,
-    String label = '',
-    bool isDefault = true,
-  }) async {
-    final model = await _source.deriveFromSeed(
-      seed: seed,
-      network: network,
-      scriptType: scriptType,
-      label: label,
-      isDefault: isDefault,
-    );
-
-    return model.toEntity();
-  }
-
-  @override
-  Future<WalletMetadata> deriveFromXpub({
-    required String xpub,
-    required Network network,
-    required ScriptType scriptType,
-    String label = '',
-  }) async {
-    final model = await _source.deriveFromXpub(
-      xpub: xpub,
-      network: network,
-      scriptType: scriptType,
-      label: label,
-    );
-
-    return model.toEntity();
-  }
-
-  @override
-  Future<void> store(WalletMetadata metadata) {
-    final model = WalletMetadataModel.fromEntity(metadata);
-    return _source.store(model);
-  }
-
-  @override
-  Future<WalletMetadata?> get(String walletId) async {
-    final model = await _source.get(walletId);
-    if (model == null) {
-      return null;
-    }
-    return model.toEntity();
-  }
-
-  @override
-  Future<List<WalletMetadata>> getAll() async {
-    final models = await _source.getAll();
-    return Future.value(models.map((e) => e.toEntity()).toList());
-  }
-
-  @override
-  Future<void> delete(String walletId) {
-    return _source.delete(walletId);
-  }
-
-  @override
-  Future<WalletMetadata> getDefault() async {
-    final all = await getAll();
-    return all.firstWhere(
-      (e) => e.isDefault,
-      orElse: () => throw "Default wallet not found",
-    );
-  }
-}
