--- lib/core/wallet/domain/repositories/liquid_wallet_repository.dart
@@ -1,1 +1,0 @@
-abstract class LiquidWalletRepository {}
