+++ lib/core/wallet/domain/usecases/get_balance_usecase.dart
@@ -1,4 +1,4 @@
-import 'package:bb_mobile/core/wallet/domain/entity/balance.dart';
+import 'package:bb_mobile/core/wallet/data/models/balance_model.dart';
 import 'package:bb_mobile/core/wallet/domain/services/wallet_manager_service.dart';
 
 /// Use case to retrieve the balance of a specific wallet
