+++ lib/features/recover_wallet/presentation/bloc/recover_wallet_bloc.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 
 import 'package:bb_mobile/core/seed/domain/usecases/find_mnemonic_words_usecase.dart';
-import 'package:bb_mobile/core/wallet/domain/entity/wallet_metadata.dart';
+import 'package:bb_mobile/core/wallet/domain/entity/wallet.dart';
 import 'package:bb_mobile/features/onboarding/domain/usecases/create_default_wallets_usecase.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
